// src/app/api/generated/model/alcance-de-serie.ts
/**
 * AKINE API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { TurnoOmitido } from './turno-omitido';
import { Turno } from './turno';


/**
 * Que turnos toca una operacion con alcance y cuales deja como estan. En la previsualizacion, lo que tocaria; en la respuesta del comando, lo que toco.
 */
export interface AlcanceDeSerie { 
    /**
     * Turnos futuros pendientes que la operacion toca. Su tamano es lo que se confirma en `cantidadConfirmada`.
     */
    afectados?: Array<Turno>;
    alcance?: AlcanceDeSerieAlcanceEnum;
    /**
     * Candidatos que NO se tocan, con su motivo
     */
    omitidos?: Array<TurnoOmitido>;
    serieId?: number;
    /**
     * Turno pivote, si se indico
     */
    turnoId?: number;
}
export enum AlcanceDeSerieAlcanceEnum {
    ESTE = 'ESTE',
    ESTE_Y_SIGUIENTES = 'ESTE_Y_SIGUIENTES',
    TODA_LA_SERIE = 'TODA_LA_SERIE'
};